backendserver/requirements.txt
fastapi==0.78.0
uvicorn==0.18.2
psycopg2-binary==2.9.3
requests==2.26.0
beautifulsoup4==4.10
pandas==1.3.4